LayoutTests/imported/w3c/web-platform-tests/intersection-observer/multiple-thresholds-expected.txt


PASS Observer with multiple thresholds. 
FAIL First rAF. assert_equals: entries.length expected 1 but got 0
FAIL document.scrollingElement.scrollTop = 120 assert_equals: entries.length expected 2 but got 0
FAIL document.scrollingElement.scrollTop = 160 assert_equals: entries.length expected 3 but got 0
FAIL document.scrollingElement.scrollTop = 200 assert_equals: entries.length expected 4 but got 0
FAIL document.scrollingElement.scrollTop = 240 assert_equals: entries.length expected 5 but got 0
FAIL document.scrollingElement.scrollTop = window.innerHeight + 140 assert_equals: entries.length expected 6 but got 0
FAIL document.scrollingElement.scrollTop = window.innerHeight + 160 assert_equals: entries.length expected 7 but got 0
FAIL document.scrollingElement.scrollTop = window.innerHeight + 200 assert_equals: entries.length expected 8 but got 0
